# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Portfolio server API documentation.",
        "title": "0xBase-Corp API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "0xSantos",
            "url": "http://www.0xbase.org",
            "email": "help@0xbase.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5050",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/bundles": {
            "post": {
                "description": "Create bundle with wallet addresses information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bundles"
                ],
                "summary": "Create bundle with wallet addresses information",
                "parameters": [
                    {
                        "description": "Bundles request",
                        "name": "addresses",
                        "in": "body",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/bundles/:id": {
            "get": {
                "description": "Fetch bundle addresses information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bundles"
                ],
                "summary": "Fetch bundle addresses information",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int",
                        "description": "Bundle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            }
        },
...
        "/portfolio/btc": {
            "get": {
                "description": "Retrieves information for a given Bitcoin Wallet address using the BTC.com API.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bitcoin"
                ],
                "summary": "Fetch Bitcoin Wallet Information",
                "parameters": [
                    {
                        "type": "array",
                        "format": "string",
                        "description": "Wallet Addresses",
                        "name": "addresses",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.PortfolioResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            }
        },
        "/portfolio/btc-wallet/{wallet_id}": {
            "get": {
                "description": "Retrieve BTC portfolio details, including BitcoinAddressInfo, for a specific wallet.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bitcoin"
                ],
                "summary": "Get BTC portfolio for a wallet",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int",
                        "description": "Wallet ID",
                        "name": "wallet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int",
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int",
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GlobalWallet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errors.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.BitcoinAddressInfo": {
            "type": "object",
            "properties": {
                "address_id": {
                    "type": "integer"
                },
                "balance": {
                    "type": "number"
                },
                "btc_asset_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "first_tx": {
                    "type": "string"
                },
                "last_tx": {
                    "type": "string"
                },
                "received": {
                    "type": "number"
                },
                "sent": {
                    "type": "number"
                },
                "tx_count": {
                    "type": "integer"
                },
                "unconfirmed_received": {
                    "type": "number"
                },
                "unconfirmed_sent": {
                    "type": "number"
                },
                "unconfirmed_tx_count": {
                    "type": "integer"
                },
                "unspent_tx_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.BitcoinBtcComV1": {
            "type": "object",
            "properties": {
                "bitcoin_address_info": {
                    "$ref": "#/definitions/models.BitcoinAddressInfo"
                },
                "btc_asset_id": {
                    "description": "Primary key",
                    "type": "integer"
                },
                "btc_usd_price": {
                    "type": "number"
                },
                "coingecko_price_feed": {
                    "$ref": "#/definitions/models.CoingeckoPriceFeed"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "wallet_id": {
                    "description": "Foreign key to global_wallets",
                    "type": "integer"
                }
            }
        },
        "models.ChainDetails": {
            "type": "object",
            "properties": {
                "chian_id": {
                    "type": "integer"
                },
                "community_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "native_token_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "usd_value": {
                    "type": "number"
                },
                "wallet_id": {
                    "type": "integer"
                },
                "wrapped_token_id": {
                    "type": "string"
                }
            }
        },
        "models.CoingeckoPriceFeed": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "description": "database id",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.EvmAssetsDebankV1": {
            "type": "object",
            "properties": {
                "chain_list_json": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "evm_asset_id": {
                    "description": "Primary key",
                    "type": "integer"
                },
                "nft_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.NFTList"
                    }
                },
                "token_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TokenList"
                    }
                },
                "total_usd_value": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "wallet_id": {
                    "description": "Foreign key to global_wallets",
                    "type": "integer"
                }
            }
        },
        "models.GlobalWallet": {
            "type": "object",
            "properties": {
                "api_endpoint": {
                    "type": "string"
                },
                "api_version": {
                    "type": "string"
                },
                "bitcoin_btc_com_v1": {
                    "$ref": "#/definitions/models.BitcoinBtcComV1"
                },
                "blockchain_type": {
                    "type": "string"
                },
                "bundle_id": {
                    "type": "integer"
                },
                "chain_details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ChainDetails"
                    }
                },
                "evm_assets_debank_v1": {
                    "$ref": "#/definitions/models.EvmAssetsDebankV1"
                },
                "last_updated_at": {
                    "type": "string"
                },
                "portfolio_id": {
                    "type": "integer"
                },
                "solana_assets_moralis_v1": {
                    "description": "relations use in json responses (optional)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.SolanaAssetsMoralisV1"
                        }
                    ]
                },
                "wallet_address": {
                    "type": "string"
                },
                "wallet_id": {
                    "type": "integer"
                }
            }
        },
        "models.NFT": {
            "type": "object",
            "properties": {
                "amount_raw": {
                    "type": "string"
                },
                "associated_token_address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "decimals": {
                    "type": "string"
                },
                "mint": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nft_id": {
                    "type": "integer"
                },
                "solana_asset_id": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.NFTList": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "chain": {
                    "type": "string"
                },
                "collection_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "content_type": {
                    "type": "string"
                },
                "contract_id": {
                    "type": "string"
                },
                "contract_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "detail_url": {
                    "type": "string"
                },
                "evm_asset_id": {
                    "type": "integer"
                },
                "id": {
                    "description": "nft id",
                    "type": "string"
                },
                "inner_id": {
                    "type": "string"
                },
                "is_erc1155": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "nft_id": {
                    "description": "database id",
                    "type": "integer"
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "total_supply": {
                    "type": "integer"
                },
                "updated_at": {
                    "description": "Attributes   json.RawMessage `gorm:\"type:jsonb\" json:\"attributes\"`\nPayToken     json.RawMessage `gorm:\"type:jsonb\" json:\"pay_token\"`",
                    "type": "string"
                },
                "usd_price": {
                    "type": "number"
                }
            }
        },
        "models.SolanaAssetsMoralisV1": {
            "type": "object",
            "properties": {
                "coingecko_price_feed": {
                    "$ref": "#/definitions/models.CoingeckoPriceFeed"
                },
                "lamports": {
                    "type": "string"
                },
                "last_updated_at": {
                    "type": "string"
                },
                "nfts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.NFT"
                    }
                },
                "solana": {
                    "type": "string"
                },
                "solana_asset_id": {
                    "type": "integer"
                },
                "tokens": {
                    "description": "relations use in json responses (optional)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Token"
                    }
                },
                "total_nfts_count": {
                    "type": "integer"
                },
                "total_tokens_count": {
                    "type": "integer"
                },
                "wallet_id": {
                    "type": "integer"
                }
            }
        },
        "models.Token": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "amount_raw": {
                    "type": "string"
                },
                "associated_token_address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "decimals": {
                    "type": "string"
                },
                "mint": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "solana_asset_id": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                },
                "token_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.TokenList": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "chain": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "decimals": {
                    "type": "integer"
                },
                "display_symbol": {
                    "type": "string"
                },
                "evm_asset_id": {
                    "type": "integer"
                },
                "id": {
                    "description": "token id",
                    "type": "string"
                },
                "is_core": {
                    "type": "boolean"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "is_wallet": {
                    "type": "boolean"
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "optimized_symbol": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "price_24h_change": {
                    "type": "number"
                },
                "protocol_id": {
                    "type": "string"
                },
                "raw_amount": {
                    "type": "number"
                },
                "raw_amount_hex_str": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "time_at": {
                    "type": "number"
                },
                "token_id": {
                    "description": "database id",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "responses.ChainsResponse": {
            "type": "object",
            "properties": {
                "asset_id": {
                    "type": "string"
                },
                "asset_percentage": {
                    "type": "number"
                },
                "asset_symbol": {
                    "type": "string"
                },
                "chain": {
                    "type": "string"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "quantity": {
                    "type": "number"
                },
                "total_price": {
                    "type": "number"
                },
                "unit_price": {
                    "type": "number"
                },
                "wallet_id": {
                    "type": "integer"
                }
            }
        },
        "responses.PortfolioResponse": {
            "type": "object",
            "properties": {
                "asset_symbol": {
                    "type": "string"
                },
                "chains_info": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.ChainsResponse"
                    }
                },
                "portfolio_percentage": {
                    "type": "number"
                },
                "quantity": {
                    "type": "number"
                },
                "total_price": {
                    "type": "number"
                },
                "unit_price": {
                    "type": "number"
                }
            }
        }
    }
}